+++ src/main/kotlin/com/krillsson/sysapi/config/StoragePoolConfiguration.kt
@@ -1,0 +1,5 @@
+package com.krillsson.sysapi.config
+
+data class StoragePoolConfiguration(
+    val enabled: Boolean = true
+)