+++ src/main/kotlin/com/krillsson/sysapi/graphql/domain/StoragePoolInfo.kt
@@ -1,0 +1,9 @@
+package com.krillsson.sysapi.graphql.domain
+
+interface StoragePoolInfo
+object StoragePoolInfoAvailable : StoragePoolInfo
+
+data class StoragePoolInfoUnavailable(
+    val reason: String,
+    val isDisabled: Boolean
+) : StoragePoolInfo